Add Day33 Work

+++ Appointment System/Appointment-System/src/main/java/com/example/Appointment/System/controller/UserAuthController.java
@@ -1,22 +1,19 @@
 package com.example.Appointment.System.controller;
 
-import com.example.Appointment.System.jwt.JwtUtils;
+import com.example.Appointment.System.jwt.utils.JwtUtils;
+import com.example.Appointment.System.jwt.utils.RequestUtils;
 import com.example.Appointment.System.model.dto.LoginDTO;
 import com.example.Appointment.System.model.dto.PasswordDTO;
 import com.example.Appointment.System.model.dto.UserDTO;
 import com.example.Appointment.System.model.entity.MUser;
-import com.example.Appointment.System.model.mapper.PatientMapper;
 import com.example.Appointment.System.model.mapper.UserMapper;
-import com.example.Appointment.System.service.Imp.PatientServiceImp;
-import com.example.Appointment.System.service.Imp.UserServiceImp;
-import com.example.Appointment.System.service.PatientService;
 import com.example.Appointment.System.service.UserService;
 import com.example.Appointment.System.service.UserValidationService;
+import com.example.Appointment.System.service.ValidationService;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Controller;
@@ -24,10 +21,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.util.HashMap;
 import java.util.Map;
-import java.util.Objects;
-import java.util.Set;
 
 @Controller
 @RequiredArgsConstructor
@@ -35,10 +29,8 @@ public class UserAuthController {
     private final UserService userService;
     private final UserMapper userMapper;
     private final JwtUtils jwtUtils;
-    private final AuthenticationManager authenticationManager;
-    private final PasswordEncoder passwordEncoder;
     private final UserValidationService userValidationService;
-
+    private final ValidationService validationService;
     @GetMapping("/")
     public String startPage(HttpServletRequest request){
         if(SecurityContextHolder.getContext().getAuthentication().isAuthenticated()){
@@ -54,20 +46,17 @@ public class UserAuthController {
     public String loginUser(){
         return "login";
     }
-
     @GetMapping("/home")
     @ResponseBody
     public ResponseEntity<UserDTO> getHome(){
-
         return ResponseEntity.ok(userMapper.toUserDTO(userService.findUserByContact(SecurityContextHolder.getContext().getAuthentication().getName())));
     }
-
     @PostMapping("/register")
     @ResponseBody
     public ResponseEntity<?> registerUser(@RequestBody UserDTO userDTO){
-       if(!userService.validateUserDetails(userDTO).isEmpty()){
-           return ResponseEntity.badRequest().body(userService.validateUserDetails(userDTO));
-       }
+        if(!validationService.validatePatientDetails(userDTO).isEmpty()){
+            return ResponseEntity.badRequest().body(validationService.validatePatientDetails(userDTO));
+        }
         return ResponseEntity.ok(
                 userMapper.toUserDTO(userService.saveUser(userMapper.toUser(userDTO)))
         );
@@ -75,32 +64,35 @@ public class UserAuthController {
     @PostMapping("/login")
     @ResponseBody
     public ResponseEntity<?> loginUser(@RequestBody LoginDTO loginDTO){
-        if(!userService.isExitUserByContact(loginDTO.getContact())){
+
+        MUser user= userService.findUserByContact(loginDTO.getContact());
+        if(user==null){
             return ResponseEntity.badRequest().body("User doesn't exit..");
         }
-        if(!passwordEncoder.matches(loginDTO.getPassword(), userService.findUserByContact(loginDTO.getContact()).getPassword())){
+
+        if(!userValidationService.isExitUserPassword(loginDTO.getContact(),loginDTO.getPassword())){
             return ResponseEntity.badRequest().body("Password is incorrect..");
         }
-        authenticationManager.authenticate(
-                new UsernamePasswordAuthenticationToken(loginDTO.getContact(),loginDTO.getPassword()));
-        MUser user= userService.findUserByContact(loginDTO.getContact());
-        String token=jwtUtils.generateToken(user.getName(),loginDTO.getContact());
-        user.setIsActive(true);
-        userService.saveUser(user);
+        String token=userService.authenticateUser(user,loginDTO);
         return ResponseEntity.ok(Map.of("token",token));
     }
     @PostMapping("/signout")
     @ResponseBody
-    public ResponseEntity<?> logoutUser(HttpServletRequest request){
-        String authHeader = request.getHeader("Authorization");
-        if (authHeader == null || !authHeader.startsWith("Bearer ")) {
-            return ResponseEntity.badRequest().body("Invalid token");
+    public ResponseEntity<?> logoutUser(){
+        HttpServletRequest request= RequestUtils.getCurrentHttpRequest();
+        if(request==null){
+            return ResponseEntity.badRequest().body("HttpServletRequest Object is empty");
         }
+        String authHeader = request.getHeader("Authorization");
         String token = authHeader.substring(7);
         String contact=jwtUtils.extractContact(token);
-        MUser user= userService.findUserByContact(contact);
-        user.setIsActive(false);
-        userService.saveUser(user);
+        if (!authHeader.startsWith("Bearer ")) {
+            return ResponseEntity.badRequest().body("Invalid token");
+        }
+        if(!userValidationService.isExitUserByContact(contact)){
+            return ResponseEntity.badRequest().body("User Contact doesn't exit");
+        }
+        userService.logoutUser(contact);
         return ResponseEntity.ok("Logout successfully");
     }
 
@@ -113,8 +105,11 @@ public class UserAuthController {
             @RequestParam String gender,
             @RequestParam(required = false) MultipartFile photo
     ) throws IOException {
-       if(!userValidationService.isValidGender(gender).isEmpty()){
-           return ResponseEntity.badRequest().body(userValidationService.isValidGender(gender));
+       if(!userValidationService.isValidGender(gender)){
+           return ResponseEntity.badRequest().body("Gender must be male,female or other");
+       }
+       if(!userValidationService.isValidEmailFormat(email)){
+           return ResponseEntity.badRequest().body("Email must contain @");
        }
         return ResponseEntity.ok(
                 userMapper.toUserDTO(userService.updatePatientWithOutPassword(
@@ -124,6 +119,9 @@ public class UserAuthController {
     @DeleteMapping("/delete/{id}")
     @ResponseBody
     public ResponseEntity<String> deleteUserById(@PathVariable Long id){
+        if(!userValidationService.isExitUserById(id)){
+            return ResponseEntity.badRequest().body("User doesn't exit!");
+        }
         userService.deleteUser(id);
          return ResponseEntity.ok("Deleted successfully");
     }
@@ -131,7 +129,7 @@ public class UserAuthController {
     @ResponseBody
     public ResponseEntity<?> fetchUser(){
         String contact=SecurityContextHolder.getContext().getAuthentication().getName();
-        if(!userService.isExitUserByContact(contact)){
+        if(!userValidationService.isExitUserByContact(contact)){
             return ResponseEntity.badRequest().body("User doesn't exit..");
         }
         return ResponseEntity.ok(userMapper.toUserDTO(userService.findUserByContact(contact)));
@@ -140,19 +138,12 @@ public class UserAuthController {
     @PostMapping("/user/change-password")
     @ResponseBody
     public ResponseEntity<?> updateUserPassword(@RequestBody PasswordDTO passwordDTO){
-
         String contact=SecurityContextHolder.getContext().getAuthentication().getName();
-        if (!passwordEncoder.matches(passwordDTO.getPassword(), userService.findUserByContact(contact).getPassword())){
-            return ResponseEntity.badRequest().body("Current Password is incorrect..");
+        if(!userValidationService.isExitUserByContact(contact)){
+            return ResponseEntity.badRequest().body("User doesn't exit..");
         }
-        if(!userValidationService.isExitUserByContact(contact).isEmpty()){
-            return ResponseEntity.badRequest().body(userValidationService.isExitUserByContact(contact));
-        }
-        if (!Objects.equals(passwordDTO.getNewPassword(), passwordDTO.getConfirmPassword())){
-            return ResponseEntity.badRequest().body("Password doesn't match..");
-        }
-        if (!userValidationService.isValidUserPasswordLength(passwordDTO.getNewPassword()).isEmpty()){
-            return ResponseEntity.badRequest().body(userValidationService.isValidUserPasswordLength(passwordDTO.getNewPassword()));
+        if(!validationService.validatePatientPasswordUpdate(passwordDTO).isEmpty()){
+            return ResponseEntity.badRequest().body(validationService.validatePatientPasswordUpdate(passwordDTO));
         }
         userService.updateUserPassword(contact,passwordDTO);
         return ResponseEntity.ok("Password updated successfully");
